# documentation/openapi.yaml
openapi: 3.0.3
info:
  title: nano-toolset API
  version: 1.0.0
  description: |
    OpenAPI contract for the plugin HTTP toolset routes.
    This API is framework-agnostic and can be consumed by humans, scripts, or autonomous agents.
    All responses use a common envelope:
    - success: boolean
    - data: object|null
    - error: { error, message }|null
servers:
  - url: http://{host}
    description: HTTP listener from HTTP_LISTEN_ADDR
    variables:
      host:
        default: 127.0.0.1:3123
externalDocs:
  description: Canonical error catalog
  url: ./errors.md

tags:
  - name: wallet
  - name: payment
  - name: credits
  - name: donate

paths:
  /wallet/balance:
    get:
      tags: [wallet]
      summary: Get wallet balance
      operationId: getWalletBalance
      responses:
        '200':
          description: Balance retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseBalance'
        '400':
          $ref: '#/components/responses/ApiError'
        '500':
          $ref: '#/components/responses/ApiError'
        '502':
          $ref: '#/components/responses/ApiError'

  /wallet/send:
    post:
      tags: [wallet]
      summary: Send Nano to a recipient address
      operationId: sendNano
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendNanoRequestApi'
      responses:
        '200':
          description: Transfer submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseSendNano'
        '400':
          $ref: '#/components/responses/ApiError'
        '500':
          $ref: '#/components/responses/ApiError'
        '502':
          $ref: '#/components/responses/ApiError'

  /payment/request:
    post:
      tags: [payment]
      summary: Create a payment request
      operationId: createPaymentRequest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentRequestApi'
      responses:
        '200':
          description: Payment request created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseCreatePayment'
        '400':
          $ref: '#/components/responses/ApiError'
        '500':
          $ref: '#/components/responses/ApiError'
        '502':
          $ref: '#/components/responses/ApiError'

  /payment/status/{transaction_id}:
    get:
      tags: [payment]
      summary: Get payment status by transaction ID
      operationId: getPaymentStatus
      parameters:
        - name: transaction_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Payment status retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponsePaymentStatus'
        '400':
          $ref: '#/components/responses/ApiError'
        '500':
          $ref: '#/components/responses/ApiError'
        '502':
          $ref: '#/components/responses/ApiError'

  /credits:
    get:
      tags: [credits]
      summary: Get available credits and price tiers
      operationId: getCredits
      responses:
        '200':
          description: Credits info retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseCredits'
        '400':
          $ref: '#/components/responses/ApiError'
        '500':
          $ref: '#/components/responses/ApiError'
        '502':
          $ref: '#/components/responses/ApiError'

  /credits/topup/{credits_amount}:
    post:
      tags: [credits]
      summary: Top up credits by predefined amount
      operationId: topupCredits
      parameters:
        - name: credits_amount
          in: path
          required: true
          schema:
            type: integer
            format: int32
            enum: [10, 50, 100, 500, 1000, 5000, 10000, 50000, 100000]
      responses:
        '200':
          description: Credits topped up
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseTopupCredits'
        '400':
          $ref: '#/components/responses/ApiError'
        '500':
          $ref: '#/components/responses/ApiError'
        '502':
          $ref: '#/components/responses/ApiError'

  /donate/{amount}:
    post:
      tags: [donate]
      summary: Donate a Nano amount
      operationId: donate
      parameters:
        - name: amount
          in: path
          required: true
          schema:
            type: string
          description: Nano amount as decimal string
      responses:
        '200':
          description: Donation attempt completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseDonate'
        '400':
          $ref: '#/components/responses/ApiError'
        '500':
          $ref: '#/components/responses/ApiError'
        '502':
          $ref: '#/components/responses/ApiError'

components:
  responses:
    ApiError:
      description: Error response envelope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponseError'

  schemas:
    ApiError:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string

    ApiResponseError:
      type: object
      required: [success, data, error]
      properties:
        success:
          type: boolean
          enum: [false]
          description: Always false for failed requests
        data:
          nullable: true
          description: Always null when success is false
        error:
          $ref: '#/components/schemas/ApiError'

    ApiResponseBalance:
      type: object
      required: [success, data, error]
      properties:
        success:
          type: boolean
          enum: [true]
          description: Always true for successful requests
        data:
          $ref: '#/components/schemas/BalanceResponseApi'
          description: Response payload
        error:
          nullable: true
          description: Always null when success is true

    ApiResponseSendNano:
      type: object
      required: [success, data, error]
      properties:
        success:
          type: boolean
          enum: [true]
          description: Always true for successful requests
        data:
          $ref: '#/components/schemas/SendNanoResponseApi'
          description: Response payload
        error:
          nullable: true
          description: Always null when success is true

    ApiResponseCreatePayment:
      type: object
      required: [success, data, error]
      properties:
        success:
          type: boolean
          enum: [true]
          description: Always true for successful requests
        data:
          $ref: '#/components/schemas/CreatePaymentResponseApi'
          description: Response payload
        error:
          nullable: true
          description: Always null when success is true

    ApiResponsePaymentStatus:
      type: object
      required: [success, data, error]
      properties:
        success:
          type: boolean
          enum: [true]
          description: Always true for successful requests
        data:
          $ref: '#/components/schemas/StatusPaymentnApi'
          description: Response payload
        error:
          nullable: true
          description: Always null when success is true

    ApiResponseCredits:
      type: object
      required: [success, data, error]
      properties:
        success:
          type: boolean
          enum: [true]
          description: Always true for successful requests
        data:
          $ref: '#/components/schemas/GetCreditsResponse'
          description: Response payload
        error:
          nullable: true
          description: Always null when success is true

    ApiResponseTopupCredits:
      type: object
      required: [success, data, error]
      properties:
        success:
          type: boolean
          enum: [true]
          description: Always true for successful requests
        data:
          $ref: '#/components/schemas/TopupCreditsSuccessResponse'
          description: Response payload
        error:
          nullable: true
          description: Always null when success is true

    ApiResponseDonate:
      type: object
      required: [success, data, error]
      properties:
        success:
          type: boolean
          enum: [true]
          description: Always true for successful requests
        data:
          $ref: '#/components/schemas/DonateResponseApi'
          description: Response payload
        error:
          nullable: true
          description: Always null when success is true

    BalanceResponseApi:
      type: object
      required: [account, balance, balance_raw, pending, pending_raw]
      properties:
        account:
          type: string
        balance:
          type: string
          description: Balance in Nano
        balance_raw:
          type: string
        pending:
          type: string
          description: Pending amount in Nano
        pending_raw:
          type: string

    SendNanoRequestApi:
      type: object
      required: [recipient_address, amount]
      properties:
        recipient_address:
          type: string
        amount:
          type: string
          description: Nano amount as decimal string

    SendNanoResponseApi:
      type: object
      required: [amount, recipient]
      properties:
        amount:
          type: string
        recipient:
          type: string

    CreatePaymentRequestApi:
      type: object
      required: [receive_address, amount]
      properties:
        receive_address:
          type: string
        amount:
          type: string
        redirect_url:
          type: string
          nullable: true

    CreatePaymentResponseApi:
      type: object
      required: [receive_address, amount, transaction_id]
      properties:
        receive_address:
          type: string
        amount:
          type: string
        transaction_id:
          type: string

    StatusPaymentnApi:
      type: object
      required: [success, message, transaction_id, is_paid]
      properties:
        success:
          type: boolean
        message:
          type: string
        transaction_id:
          type: string
        is_paid:
          type: boolean

    GetCreditsResponse:
      type: object
      required:
        - credits
        - current_credits_price_10
        - current_credits_price_50
        - current_credits_price_100
        - current_credits_price_500
        - current_credits_price_1000
        - current_credits_price_5000
        - current_credits_price_10000
        - current_credits_price_50000
        - current_credits_price_100000
      properties:
        credits:
          type: integer
          format: int32
        current_credits_price_10:
          type: number
          format: double
        current_credits_price_50:
          type: number
          format: double
        current_credits_price_100:
          type: number
          format: double
        current_credits_price_500:
          type: number
          format: double
        current_credits_price_1000:
          type: number
          format: double
        current_credits_price_5000:
          type: number
          format: double
        current_credits_price_10000:
          type: number
          format: double
        current_credits_price_50000:
          type: number
          format: double
        current_credits_price_100000:
          type: number
          format: double

    TopupCreditsSuccessResponse:
      type: object
      required: [topped_up_credits, new_credits_balance]
      properties:
        topped_up_credits:
          type: integer
          format: int32
        new_credits_balance:
          type: integer
          format: int32

    DonateResponseApi:
      type: object
      required: [success, amount, message]
      properties:
        success:
          type: boolean
        amount:
          type: string
        message:
          type: string